Handle merge conflicts
